requirements.txt
certifi==2024.8.30
Flask==3.0.3
Flask_Cors==5.0.0
flask_wtf==1.2.1
pymongo==4.10.1
python-dotenv==1.0.1
Requests==2.32.3
Werkzeug==3.0.4
WTForms==3.1.2